CRUD Spring Boot

+++ CalcFx/src/main/java/pe/edu/upeu/calcfx/servicio/CalcServiceI.java
@@ -7,8 +7,8 @@ public interface CalcServiceI {
 
     public void guardarResultados(CalcTO to);//C
     public List<CalcTO> obtenerResultados();//R
-    public void actualizarResultados(CalcTO to, int index);//U
-    public void eliminarResultados(int index);//D
+    public void actualizarResultados(CalcTO to, Long index);//U
+    public void eliminarResultados(Long index);//D
 
 
 }
